split codebase
