+++ src/util.rs
@@ -1,0 +1,5 @@
+use std::fs::remove_file;
+
+pub fn clean() {
+    let _ignored = remove_file("temp");
+}
